% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ids.R
\name{collapsed_entrezg_to_gsymbol}
\alias{collapsed_entrezg_to_gsymbol}
\title{Map collapsed entrezg ids to collapsed gene symbols}
\usage{
collapsed_entrezg_to_gsymbol(x, org.xx.eg.db, split = ";")
}
\arguments{
\item{x}{vector with collapsed entrezg ids}

\item{org.xx.eg.db}{respective annotation map}

\item{split}{collapsing character to split on}
}
\value{
vector with collapsed gene symbols
}
\description{
Map collapsed entrezg ids to collapsed gene symbols
}
\examples{
\dontrun{
   collapsed_entrezg_to_gsymbol('1;2;3', org.Hs.eg.db::org.Hs.eg.db)
}
}
