% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Constants}
\alias{Constants}
\alias{ANNOTATED_ORGANISMS}
\alias{PANTHER_ORGANISMS}
\alias{GO_COLLECTIONS}
\alias{GENE_SET_COLLECTIONS}
\title{Constants in autonomics.ora}
\format{
An object of class \code{list} of length 7.

An object of class \code{character} of length 6.

An object of class \code{character} of length 3.

An object of class \code{character} of length 4.
}
\usage{
ANNOTATED_ORGANISMS

PANTHER_ORGANISMS

GO_COLLECTIONS

GENE_SET_COLLECTIONS
}
\description{
Which organisms have available annotations?
Which GO collections are supported?
Which gene set collections are supported?
}
\examples{
ANNOTATED_ORGANISMS
PANTHER_ORGANISMS
GO_COLLECTIONS
GENE_SET_COLLECTIONS
}
\keyword{datasets}
