% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_ora_detectome_in_genome}
\alias{default_ora_detectome_in_genome}
\title{Default value for argument detectome_against_genome}
\usage{
default_ora_detectome_in_genome()
}
\value{
default value of detectome_against_genome
}
\description{
Should ora of detectome versus genome be performed?
}
