% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_universe}
\alias{default_universe}
\title{default universe}
\usage{
default_universe(object)
}
\arguments{
\item{object}{eset}
}
\value{
default value of universe
}
\description{
default universe
}
