% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_feature_sets.R
\name{extract_ora_sets}
\alias{extract_ora_sets}
\title{Extract detectome go/interpro sets}
\usage{
extract_ora_sets(object, oraset_var)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{oraset_var}{ora set variable}
}
\value{
list (names = setids, values = fids)
}
\description{
Extract detectome go/interpro sets
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   object \%>\% autonomics.ora::extract_ora_sets('goid')     \%>\% head(2)
   object \%>\% autonomics.ora::extract_ora_sets('interpro') \%>\% head(2)
}
}
