% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_feature_sets.R
\name{extract_ora_universe}
\alias{extract_ora_universe}
\title{Extract ora universe}
\usage{
extract_ora_universe(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
character vector
}
\description{
Extract ora universe
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   object \%>\% autonomics.ora::extract_ora_universe()       \%>\% head(2)
}
}
