% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ora_on_eset.R
\name{get_entrezg}
\alias{get_entrezg}
\title{Get entrezg from eset}
\usage{
get_entrezg(object, load_org.xx.xx.db)
}
\arguments{
\item{object}{eset}

\item{load_org.xx.xx.db}{load_org.xx.xx.db}
}
\value{
vector with entrez gene ids
}
\description{
Get entrezg from eset
}
\examples{
require(magrittr)
if (require(autonomics.data) & require(org.Hs.eg.db)){
   object <- autonomics.data::billing2016[1:10, ]
   object \%>\% get_entrezg(org.Hs.eg.db::org.Hs.eg.db)
}
if (require(atkin.2014) & require(org.Hs.eg.db)){
   atkin.2014::soma[1:10, ] \%>\%
   get_entrezg(org.Hs.eg.db::org.Hs.eg.db)
}
}
