% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ora_on_eset.R
\name{get_entrezg_from_gsymbols}
\alias{get_entrezg_from_gsymbols}
\title{Get entrez from gene symbols}
\usage{
get_entrezg_from_gsymbols(object, load_org.xx.xx.db)
}
\arguments{
\item{object}{eset}

\item{load_org.xx.xx.db}{load_org.xx.xx.db}
}
\value{
vector with entrez gene ids
}
\description{
Get entrez from gene symbols
}
\examples{
require(magrittr)
if (require(autonomics.data) & require(org.Hs.eg.db)){
   object <- autonomics.data::billing2016
   load_org.xx.xx.db <- org.Hs.eg.db::org.Hs.eg.db
   object \%>\% get_entrezg_from_gsymbols(load_org.xx.xx.db)
}
if (require(atkin.2014) & require(org.Hs.eg.db)){
   atkin.2014::soma[1:10, ] \%>\%
   get_entrezg_from_gsymbols(org.Hs.eg.db::org.Hs.eg.db)
}
}
