% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ora_on_eset.R
\name{get_entrezg_from_uniprot}
\alias{get_entrezg_from_uniprot}
\title{Get entrez gene ids from uniprot}
\usage{
get_entrezg_from_uniprot(object, load_org.xx.xx.db)
}
\arguments{
\item{object}{eset}

\item{load_org.xx.xx.db}{load_org.xx.xx.db}
}
\value{
entrezg ids
}
\description{
Get entrez gene ids from uniprot
}
\examples{
require(magrittr)
if (require(autonomics.data) & require(org.Hs.eg.db)){
   object <- autonomics.data::billing2016
   object[1:10, ] \%>\% get_entrezg_from_uniprot(org.Hs.eg.db::org.Hs.eg.db)
}
if (require(atkin.2014) & require(org.Hs.eg.db)){
   object <- atkin.2014::soma
   object[1:10, ] \%>\% get_entrezg_from_uniprot(org.Hs.eg.db::org.Hs.eg.db)
}
}
