% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ids.R
\name{gsymbol_set_to_entrezg_set}
\alias{gsymbol_set_to_entrezg_set}
\title{Map gene symbols to entrezg ids}
\usage{
gsymbol_set_to_entrezg_set(x, annotation_map)
}
\arguments{
\item{x}{vector with gene symbols}

\item{annotation_map}{org.xx.eg.db object}
}
\value{
vector with entrezg ids
}
\description{
Map gene symbols to entrezg ids
}
\examples{
\dontrun{
   gsymbol_set_to_entrezg_set(c("A1BG", "A2M"), org.Hs.eg.db::org.Hs.eg.db)
}
}
