% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_organism.R
\name{infer_organism}
\alias{infer_organism}
\title{Infer organism from feature annotations}
\usage{
infer_organism(object)
}
\arguments{
\item{object}{eSet}
}
\value{
A string corresponding to a name from \code{\link{ANNOTATED_ORGANISMS}}
}
\description{
Infer organism from feature annotations
}
\examples{
   library(magrittr)
   if (require(autonomics.data)){
     autonomics.data::billing2016 \%>\% infer_organism()
   }
   if (require(atkin.2014)){
      atkin.2014::soma \%>\% infer_organism()
   }
   if (require(billing.differentiation.data)){
      require(magrittr)
      billing.differentiation.data::protein.ratios \%>\% infer_organism()
   }
}
