% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_organism.R
\name{infer_organism_from_uniprot_accessions}
\alias{infer_organism_from_uniprot_accessions}
\title{Infer organism from uniprot accessions}
\usage{
infer_organism_from_uniprot_accessions(object)
}
\arguments{
\item{object}{eset}
}
\description{
Infer organism from uniprot accessions
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   object \%>\% infer_organism_from_uniprot_accessions()
   autonomics.data::ALL \%>\% infer_organism_from_uniprot_accessions()
}
if (require(billing.differentiation.data)){
   object <- billing.differentiation.data::protein.ratios
   object \%>\% infer_organism_from_uniprot_accessions()
}
if (require(atkin.2014)){
   object <- atkin.2014::soma
   object \%>\% infer_organism_from_uniprot_accessions()
}
}
