% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_collection.R
\name{load_collection}
\alias{load_collection}
\title{Get pathways}
\usage{
load_collection(
  collection,
  organism,
  universe = NULL,
  min_set_size = 5,
  max_set_size = 2000
)
}
\arguments{
\item{collection}{value from GENE_SET_COLLECTIONS}

\item{organism}{value from ANNOTATED_ORGANISMS}

\item{universe}{either NULL or a vector with feature IDs defining the universe}

\item{min_set_size}{minimum number of features in a set}

\item{max_set_size}{maximum number of features in a set}
}
\description{
Get pathways
}
