% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_collection.R
\name{load_go_collection}
\alias{load_go_collection}
\title{Load GO collection}
\usage{
load_go_collection(collection_name, organism)
}
\arguments{
\item{collection_name}{value in GO_COLLECTIONS}

\item{organism}{value in ANNOTATED_ORGANISMS}
}
\value{
Named list: names are GO IDs, each element contains a vector of entrez gene IDs
}
\description{
Load GO collection
}
\examples{
gobp <- load_go_collection('GOBP', 'Homo sapiens')
}
