% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_orgdb.R
\name{load_org.xx.xx.db}
\alias{load_org.xx.xx.db}
\title{Load annotation map}
\usage{
load_org.xx.xx.db(organism)
}
\arguments{
\item{organism}{value in ANNOTATED_ORGANISMS}
}
\description{
Load annotation map
}
\examples{
load_org.xx.xx.db(organism = 'Homo sapiens')
load_org.xx.xx.db(organism = 'Xenopus laevis')
}
