% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ids.R
\name{map_collapsed_ids}
\alias{map_collapsed_ids}
\title{Map collapsed ids from one identifer format to another}
\usage{
map_collapsed_ids(x, from = c("SYMBOL", "ENTREZID", "UNIPROT"),
  to = c("SYMBOL", "ENTREZID", "UNIPROT"), org.xx.eg.db, split = ";")
}
\arguments{
\item{x}{vector with collapsed entrezg ids}

\item{from}{A string naming the type of ID.}

\item{to}{A string naming the type of ID.}

\item{org.xx.eg.db}{respective annotation map}

\item{split}{collapsing character to split on}
}
\value{
vector with collapsed gene symbols
}
\description{
Map collapsed ids from one identifer format to another
}
\examples{
\dontrun{
   map_collapsed_ids('1;2;3', from = 'ENTREZID', to = 'SYMBOL', 
                     org.xx.eg.db = org.Hs.eg.db::org.Hs.eg.db)
}
}
