% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ids.R
\name{map_ids}
\alias{map_ids}
\title{Map ids}
\usage{
map_ids(x, from, to, annotation_map)
}
\arguments{
\item{x}{vector with uniprot accessions / gene symbols}

\item{from}{'UNIPROT', 'SYMBOL', 'ENTREZID'}

\item{to}{'UNIPROT', 'SYMBOL', 'ENTREZID'}

\item{annotation_map}{org.xx.eg.db object}
}
\description{
Map a set of ids to another set of ids
}
\examples{
library(autonomics.ora)
\dontrun{
   annmap <- org.Hs.eg.db::org.Hs.eg.db
   map_ids(c('1', '2'),                     from = 'ENTREZID', to = 'SYMBOL',   annmap)
   map_ids(c('1', '2'),                     from = 'ENTREZID', to = 'UNIPROT',  annmap)
   map_ids(c("P04217", "V9HWD8", "P01023"), from = 'UNIPROT',  to = 'ENTREZID', annmap)
   map_ids(c("A1BG", "A2M"),                from = 'SYMBOL',   to = 'ENTREZID', annmap)
}
}
