% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_panther_id.R
\name{map_panther_id_to_term}
\alias{map_panther_id_to_term}
\title{Map panther class_id to class_term}
\usage{
map_panther_id_to_term(class_id)
}
\arguments{
\item{class_id}{panther class id}
}
\value{
vector with uniprots
}
\description{
Map panther class_id to class_term
}
\examples{
map_panther_id_to_term('PC00193')
map_panther_id_to_term('PC00137')
}
