% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_panther_id.R
\name{map_panther_id_to_uniprots}
\alias{map_panther_id_to_uniprots}
\title{Map panther class_id to uniprots}
\usage{
map_panther_id_to_uniprots(class_id, organism)
}
\arguments{
\item{class_id}{panther class id}

\item{organism}{organism name, formatted as 'Homo sapiens'. names(PANTHER_ORGANISMS) shows currently supported values.}
}
\value{
vector with uniprots
}
\description{
Map panther class_id to uniprots
}
\examples{
map_panther_id_to_uniprots(class_id = 'PC00193', organism = 'Homo sapiens')
map_panther_id_to_uniprots(class_id = 'PC00137', organism = 'Homo sapiens')
}
