% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ora.R
\name{run_ora}
\alias{run_ora}
\title{Run over representation analysis}
\usage{
run_ora(
  query,
  universe,
  pathway_list,
  min_set_size = 5,
  return_only_significant = FALSE,
  collapse_features_column = TRUE
)
}
\arguments{
\item{query}{Character vector of feature IDs in the selection.}

\item{universe}{Character vector of feature IDs in the whole dataset (all 
detected features).}

\item{pathway_list}{A named list of character vectors. Names correspond to 
pathways. Each component of the list is a character vector of feature IDs.}

\item{min_set_size}{An integer. The minimum number of universe genes 
required for a pathway to be included in analysis.}

\item{return_only_significant}{A logical value. Should returned results be 
limited to those that are significant?}

\item{collapse_features_column}{Logical.  Should the \code{features} column 
be collapsed into a character vector?  See return value section.}
}
\value{
A data frame with the following columns:
\describe{
  \item{rank}{Integer, 1 to the number of rows.  The rank of the significance 
  of the pathway.}
  \item{p}{Numeric, between 0 and 1. The p-value scoring the significance of
  the pathway.}
  \item{pathway}{Character. The ID of the pathway, taken from 
  \code{names(pathway_list)}.}
  \item{n_total}{Integer, positive. The number of features found in the 
  \code{pathway_list} with this pathway.}
  \item{n_detected}{Integer, positive. The number of features found in the 
  \code{universe} with this pathway.}
  \item{n_selected}{Integer, positive. The number of features found in the 
  \code{query} with this pathway.}
  \item{features}{Character or list of character.  If 
  \code{collapse_features_column} is \code{TRUE}, a character vector of semi-
  colon separated feature names.  Otherwise, a list of character vectors of 
  feature names.}
}
}
\description{
Each pathway is analyzed for an enrichment in query genes, 
using a right-sided (over representation) Fisher Exact test.
}
\details{
First define:
   q = # query features in pathway
   m = # pathway genes
   n = # non pathway genes
   k = # query genes

The right-sided Fisher exact test investigates whether:
   H0: q = random draw from universe
   H1: q > random draw from universe

Under H0:
   q ~ phyper(m, n, k)
   
So the Fisher exact test examines whether 
   phyper(q, m, n, k, lower.tail = FALSE) <= alpha
}
\note{
1. My original implementation used 'fisher.test', but that was not vectorizable.
}
