% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ora_on_eset.R
\name{run_ora_on_eset}
\alias{run_ora_on_eset}
\title{Run over representation analysis}
\usage{
run_ora_on_eset(
  object,
  contrasts = autonomics.find::default_contrasts(object),
  result_dir,
  gene_set_collections = GENE_SET_COLLECTIONS,
  top_definition = autonomics.find::default_top_definition(object),
  universe = default_universe(object),
  min_set_size = 5,
  max_set_size = Inf,
  ora_detectome_in_genome = default_ora_detectome_in_genome()
)
}
\arguments{
\item{object}{eset as returned by \code{\link[autonomics.find]{add_limma_to_fdata}}}

\item{contrasts}{contrasts vector}

\item{result_dir}{directory to which to write results}

\item{gene_set_collections}{gene set collections used for analysis (see \code{\link{GENE_SET_COLLECTIONS}} for implemented ones)}

\item{top_definition}{definition of 'top features'.}

\item{universe}{'detectome' (all detected features), 'genome' (all known features),}

\item{min_set_size}{minimum set size}

\item{max_set_size}{maximum set size}

\item{ora_detectome_in_genome}{whether to perform a detectome against genome ora (logical)}
}
\description{
Run over representation analysis
}
