% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ora_on_eset.R
\name{run_ora_on_sumexp}
\alias{run_ora_on_sumexp}
\title{Run ora on sumexp}
\usage{
run_ora_on_sumexp(
  object,
  top_definition = autonomics.find::default_top_definition(object),
  oraset_var = "goid"
)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{top_definition}{definition of 'top_features'}

\item{oraset_var}{ora set variable}
}
\description{
Run ora on sumexp
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::billing2016
   autonomics.import::fvars(object)
}
}
