% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ora.R
\name{uncollapse_and_unite}
\alias{uncollapse_and_unite}
\title{Uncollapse and unite}
\usage{
uncollapse_and_unite(x, sep = ";")
}
\arguments{
\item{x}{collapsed string}

\item{sep}{separator to strsplit on}
}
\description{
Uncollapse and unite
}
