% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ids.R
\name{uniprot_set_to_entrezg_set}
\alias{uniprot_set_to_entrezg_set}
\title{Map uniprot accessions to entrezg ids}
\usage{
uniprot_set_to_entrezg_set(x, annotation_map)
}
\arguments{
\item{x}{vector with uniprot accessions}

\item{annotation_map}{org.xx.eg.db object}
}
\value{
vector with entrezg ids
}
\description{
Map uniprot accessions to entrezg ids
}
\examples{
\dontrun{
   uniprot_set_to_entrezg_set(c("P04217", "V9HWD8", "P01023"), org.Hs.eg.db::org.Hs.eg.db)
}
}
